backend/requirements.txt
fastapi==0.109.2
uvicorn==0.27.1
python-multipart==0.0.9
pillow==10.2.0
numpy==1.26.3
pydantic==2.5.3
opencv-python==4.9.0.80
torch==2.1.2
torchvision==0.17.1
transformers==4.37.2